Add `ballFahre` method in BohleBots
Added a new method in the `BohleBots` class called `ballFahre`. This method interprets the ball direction from the method `ballRichtung` and converts it into the driving direction needed for the `fahre` method. This change allows more accurate control of the bot in relation to the position of the ball. Method usage is demonstrated in the `main.cpp`.