Refactor LED color assignment and execution mode logic
Changes are focused on refactoring the direction logic of the BohleBots and simplifying the execution mode switching procedure. LED colors are now assigned using integers, and a LED method has been implemented in the GameModes class, improving visual feedback. Additionally, the namespace 'sarcat' is no longer used, simplifying call-to-functions and providing a more straightforward coding style.